refactor(memory): move the scheduler wire types to the wire module
Moves five scheduler types from `v2/engine.ts` into `v2.ts`, where the
other wire types live: `SchedulerActionKind`,
`SchedulerObservationTransactionKind`, `SchedulerObservationAddress`,
`CompleteActionScopeSummary`, and `SchedulerActionObservation`.
`engine.ts` re-exports them next to the `SchedulerExecutionContextKey`
re-export it already had, so no consumer imports change.

`v2.ts` is the base module -- `engine.ts` imports from it and nothing
flows back -- so a wire field cannot reference a type living downstream
without an import cycle. These five ride on commits and snapshot results
across the memory boundary, so they belong on the wire side regardless.
This is a prerequisite for typing `schedulerObservation` / `observation`,
which are still `unknown`.

`ResolvedSchedulerObservationAddress` and `SchedulerWriteAddress` stay in
`engine.ts`: engine-internal, not wire shapes.

Also documents, at both declarations, that `SchedulerActionObservation`
exists twice -- here and in
`runner/src/scheduler/persistent-observation.ts` -- with each comment
naming the other's location, the two differences (address type, `branch`),
and the reason drift goes unnoticed: the wire fields carrying an
observation are `unknown`, so neither declaration constrains the other.

No runtime change.
